requirements.txt
langgraph>=0.2.0
anthropic>=0.40.0
python-docx>=1.1.0
openai>=1.0.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
pytest>=7.0.0
